+++ R/equation.R
@@ -262,19 +262,20 @@ mlxTxt <- function(file, retFile=FALSE) {
   })
   if (!retFile) .mlxtranIni()
   .exit <- FALSE
+  .monolixLib <- NULL
   if (length(file) > 1L) {
     .lines <- file
     .dirn <- getwd()
   } else {
-    if (requireNamespace("lixoftConnectors", quietly = TRUE)) {
-      if (!checkmate::testCharacter(file, min.chars = 5, len=1)) {
+    if (monolix2rxlixoftConnectors()) {
+      if (checkmate::testCharacter(file, min.chars = 5, len=1)) {
         .pre <- substr(file, 1, 4)
       } else {
         .pre <- ""
       }
       if (.pre == "lib:") {
         if (is.na(.monolix2rx$lixoftConnectors)) {
-          x <- try(lixoftConnectors::initializeLixoftConnectors(software = "monolix", force=TRUE), silent=TRUE)
+          x <- try(monolix2rxInitializeLixoftConnectors(software = "monolix", force=TRUE), silent=TRUE)
           if (inherits(x, "try-error")) {
             warning("lixoftConnectors cannot be initialized",
                     call.=FALSE)
@@ -284,19 +285,24 @@ mlxTxt <- function(file, retFile=FALSE) {
           }
         }
         if (.monolix2rx$lixoftConnectors) {
-          .ret <- try(lixoftConnectors::getLibraryModelContent(, print=FALSE), silent=TRUE)
+          .ret <- try(monolix2rxGetLibraryModelContent(file), silent=TRUE)
           if (!inherits(.ret, "try-error")) {
-            return(as.character(.ret))
+            .monolixLib <- strsplit(as.character(.ret), "\n")[[1]]
           }
         }
       }
     }
-    .f <- .mlxtranLib(file)
-    if (checkmate::testFileExists(.f, "r")) {
-      .lines <- suppressWarnings(readLines(.f))
-      .dirn <- dirname(.f)
+    if (!is.null(.monolixLib)) {
+      .lines <- .monolixLib
+      .dirn <- NULL
     } else {
-      .exit <- TRUE
+      .f <- .mlxtranLib(file)
+      if (checkmate::testFileExists(.f, "r")) {
+        .lines <- suppressWarnings(readLines(.f))
+        .dirn <- dirname(.f)
+      } else {
+        .exit <- TRUE
+      }
     }
   }
   if (!.exit) {