+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Updated for new solving option in rxode2 4.0 (and depend on the packages)
 
+* Bug fixes for importing models from `lixoftConnectors`.
+
 # monolix2rx 0.0.4
 
 * Added `ignoreline` support #22